--- lib/drives/pages/create_drive_page.dart
@@ -1,24 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_app/drives/forms/create_drive_form.dart';
-
-class CreateDrivePage extends StatefulWidget {
-  const CreateDrivePage({super.key});
-
-  @override
-  State<CreateDrivePage> createState() => _CreateDrivePageState();
-}
-
-class _CreateDrivePageState extends State<CreateDrivePage> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('Create Drive'),
-      ),
-      body: const Padding(
-        padding: EdgeInsets.symmetric(horizontal: 10, vertical: 20),
-        child: SingleChildScrollView(child: CreateDriveForm()),
-      ),
-    );
-  }
-}
