+++ lib/main.dart
@@ -1,7 +1,11 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_app/app.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:supabase_flutter/supabase_flutter.dart';
+
+import 'pages/drives_page.dart';
+import 'pages/home_page.dart';
+import 'pages/rides_page.dart';
+import 'pages/settings_page.dart';
 
 void main() async {
   await dotenv.load();
@@ -15,18 +19,67 @@ void main() async {
   runApp(const MotisApp());
 }
 
-class MotisApp extends StatelessWidget {
+class MotisApp extends StatefulWidget {
   const MotisApp({super.key});
+
+  @override
+  State<MotisApp> createState() => _MotisAppState();
+}
+
+class _MotisAppState extends State<MotisApp> {
+  int _selectedIndex = 0;
+
+  static const List<Widget> _pages = [
+    HomePage(),
+    DrivesPage(),
+    RidesPage(),
+    SettingsPage(),
+  ];
 
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      title: 'Motis Mitfahr-App',
-      debugShowCheckedModeBanner: false,
-      theme: ThemeData(
-        primarySwatch: Colors.blue,
-      ),
-      home: const App(),
-    );
+        title: 'Motis Mitfahr-App',
+        debugShowCheckedModeBanner: false,
+        theme: ThemeData(
+          primarySwatch: Colors.blue,
+        ),
+        home: Scaffold(
+          appBar: AppBar(
+            title: const Text('Motis Mitfahr-App'),
+          ),
+          body: IndexedStack(
+            index: _selectedIndex,
+            children: _pages,
+          ),
+          bottomNavigationBar: BottomNavigationBar(
+            type: BottomNavigationBarType.fixed,
+            items: const <BottomNavigationBarItem>[
+              BottomNavigationBarItem(
+                icon: Icon(Icons.home),
+                label: 'Home',
+              ),
+              BottomNavigationBarItem(
+                icon: Icon(Icons.drive_eta),
+                label: 'Drives',
+              ),
+              BottomNavigationBarItem(
+                icon: Icon(Icons.chair),
+                label: 'Rides',
+              ),
+              BottomNavigationBarItem(
+                icon: Icon(Icons.settings),
+                label: 'Settings',
+              ),
+            ],
+            currentIndex: _selectedIndex,
+            selectedItemColor: Colors.blue,
+            onTap: (index) {
+              setState(() {
+                _selectedIndex = index;
+              });
+            },
+          ),
+        ));
   }
 }
