+++ lib/models/drive.dart
@@ -1,6 +1,4 @@
-import 'package:flutter_app/util/supabase.dart';
-
-import '../../util/model.dart';
+import 'model.dart';
 
 class Drive extends Model {
   final String start;
@@ -36,10 +34,8 @@ class Drive extends Model {
     );
   }
 
-  static List<Drive> fromJsonList(List<dynamic> jsonList) {
-    return jsonList
-        .map((json) => Drive.fromJson(json as Map<String, dynamic>))
-        .toList();
+  static List<Drive> fromJsonList(List<Map<String, dynamic>> jsonList) {
+    return jsonList.map((json) => Drive.fromJson(json)).toList();
   }
 
   Map<String, dynamic> toJson() {
@@ -60,19 +56,5 @@ class Drive extends Model {
   @override
   String toString() {
     return 'Drive{id: $id, from: $start at $startTime, to: $end at $endTime, by: $driverId}';
-  }
-
-  static Future<Drive?> driveOfUserAtTime(
-      DateTime start, DateTime end, int userId) async {
-    //get all drives of user
-    final List<Drive> drives = Drive.fromJsonList(
-        await supabaseClient.from('drives').select().eq('driver_id', userId));
-    //check if drive overlaps with start and end
-    for (Drive drive in drives) {
-      if (drive.startTime.isBefore(end) && drive.endTime.isAfter(start)) {
-        return drive;
-      }
-    }
-    return null;
   }
 }
