+++ lib/pages/drives_page.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+class DrivesPage extends StatefulWidget {
+  const DrivesPage({super.key});
+
+  @override
+  State<DrivesPage> createState() => _DrivesPageState();
+}
+
+class _DrivesPageState extends State<DrivesPage> {
+  @override
+  Widget build(BuildContext context) {
+    return const Center(
+      child: Text('Drives'),
+    );
+  }
+}
