+++ lib/pages/home_page.dart
@@ -10,13 +10,8 @@ class HomePage extends StatefulWidget {
 class _HomePageState extends State<HomePage> {
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('Home'),
-      ),
-      body: const Center(
-        child: Text('Home'),
-      ),
+    return const Center(
+      child: Text('Home'),
     );
   }
 }
