+++ lib/pages/rides_page.dart
@@ -10,13 +10,8 @@ class RidesPage extends StatefulWidget {
 class _RidesPageState extends State<RidesPage> {
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('Rides'),
-      ),
-      body: const Center(
-        child: Text('Rides'),
-      ),
+    return const Center(
+      child: Text('Rides'),
     );
   }
 }
