+++ lib/pages/settings_page.dart
@@ -10,13 +10,8 @@ class SettingsPage extends StatefulWidget {
 class _SettingsPageState extends State<SettingsPage> {
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('Settings'),
-      ),
-      body: const Center(
-        child: Text('Settings'),
-      ),
+    return const Center(
+      child: Text('Settings'),
     );
   }
 }
