--- lib/util/submit_button.dart
@@ -1,27 +1,0 @@
-import 'package:flutter/material.dart';
-
-class SubmitButton extends StatelessWidget {
-  final String text;
-  final Function()? onPressed;
-
-  const SubmitButton({super.key, required this.text, this.onPressed});
-
-  @override
-  Widget build(BuildContext context) {
-    return SizedBox(
-      width: 170,
-      height: 53,
-      child: MaterialButton(
-        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(100)),
-        color: Theme.of(context).colorScheme.primary,
-        onPressed: onPressed,
-        child: Text(
-          text,
-          style: TextStyle(
-              color: Theme.of(context).colorScheme.onPrimary,
-              fontWeight: FontWeight.w500),
-        ),
-      ),
-    );
-  }
-}
